+++ src/database/mod.rs
@@ -8,13 +8,11 @@ pub mod schema;
 
 use crate::conf::env_setting;
 use crate::diesel::Connection;
-use diesel::prelude::*;
 use diesel::MysqlConnection;
 use dotenv::dotenv;
 use rocket::config::Value;
 use rocket_contrib::databases::diesel;
 use std::collections::HashMap;
-use std::env;
 
 // --------------------- DB OBJECT --------------------------------------------
 
@@ -23,6 +21,10 @@ pub struct DBConnection(diesel::MysqlConnection);
 
 // --------------------- FUNCTIONS --------------------------------------------
 
+/// Create a HashMap that can be used to feed a `Rocket::Config` object with
+/// database related information, such as the url of the used database.
+///
+/// The application database shared name is `mariadb_pool`.
 pub fn pool(database_url: &str) -> HashMap<&str, Value> {
     let mut database_config = HashMap::new();
     let mut databases = HashMap::new();
@@ -35,9 +37,6 @@ pub fn pool(database_url: &str) -> HashMap<&str, Value> {
 // --------------------- FUNCTIONS --------------------------------------------
 
 /// Get the database url set in the `.env` file.
-///
-/// The `cfg()` tricks allows us to use a different database for live & test.
-#[cfg(not(test))]
 pub fn url() -> String {
     dotenv().ok();
 
@@ -56,34 +55,12 @@ pub fn url() -> String {
     )
 }
 
-#[cfg(test)]
-pub fn url() -> String {
-    dotenv().ok();
-
-    // DB settings
-    let db_host = env_setting("TEST_DB_HOST");
-    let db_port = env_setting("TEST_DB_PORT");
-    let db_adapter = env_setting("TEST_DB_ADAPTER");
-    let db_user = env_setting("TEST_DB_USER");
-    let db_password = env_setting("TEST_DB_PASSWORD");
-    let db_database = env_setting("TEST_DB_DATABASE");
-
-    // DB url
-    format!(
-        "{}://{}:{}@{}:{}/{}",
-        db_adapter, db_user, db_password, db_host, db_port, db_database
-    )
-}
-
 /// Get a database connection.
 ///
 /// Often used when the operations are not derived from a `Rocket::Request`.
-/// If it's possible, always prefer the use of the `MyDbConn` which can be
+/// If it's possible, always prefer the use of the `DBConnection` which can be
 /// retrieved from a request.
-pub fn connection() -> MysqlConnection {
-    dotenv().ok();
-
-    let database_url = url();
+pub fn connection(database_url: &str) -> MysqlConnection {
     MysqlConnection::establish(&database_url)
         .expect(&format!("Error connecting to {}", database_url))
 }