+++ changelog.md
@@ -10,6 +10,7 @@ Summary of implemented features classified by release (github tag)
 - add check_email documentation test
 - improve documentation
 - model & db refactoring
+- write all auth related tests
 
 ## [0.1.0] - 2019-03-02